sorted tests and delete test_main.py
